requirements.txt
PyQt6>=6.5
opencv-python>=4.8
Pillow>=10.0
PyYAML>=6.0
pytest>=7.0
